+++ src/AngleSharp/Css/Dom/Internal/TypeSelector.cs
@@ -1,4 +1,4 @@
-﻿namespace AngleSharp.Css.Dom
+namespace AngleSharp.Css.Dom
 {
     using AngleSharp.Dom;
     using AngleSharp.Text;
@@ -17,14 +17,8 @@
 
         public String Text => _type;
 
-        public void Accept(ISelectorVisitor visitor)
-        {
-            visitor.Type(_type);
-        }
+        public void Accept(ISelectorVisitor visitor) => visitor.Type(_type);
 
-        public Boolean Match(IElement element, IElement scope)
-        {
-            return _type.Isi(element.LocalName);
-        }
+        public Boolean Match(IElement element, IElement scope) => _type.Isi(element.LocalName);
     }
 }