+++ src/AngleSharp/Dom/ElementExtensions.cs
@@ -203,10 +203,7 @@ namespace AngleSharp.Dom
         /// <param name="el">The element that is connected to a doc.</param>
         /// <param name="prefix">The prefix to lookup.</param>
         /// <returns>The namespace url for the prefix.</returns>
-        public static String GetCssNamespace(this IElement el, String prefix)
-        {
-            return el.Owner?.StyleSheets.LocateNamespace(prefix) ?? el.LocateNamespaceFor(prefix);
-        }
+        public static String GetCssNamespace(this IElement el, String prefix) => el.Owner?.StyleSheets.LocateNamespace(prefix) ?? el.LocateNamespaceFor(prefix);
 
         /// <summary>
         /// Checks if the element is currently hovered.
@@ -314,25 +311,25 @@ namespace AngleSharp.Dom
                 var href = element.GetAttribute(null, AttributeNames.Href);
                 return !String.IsNullOrEmpty(href);
             }
-            else if (element is HtmlButtonElement)
+            else if (element is HtmlButtonElement buttonElement)
             {
-                return !((HtmlButtonElement)element).IsDisabled;
+                return !buttonElement.IsDisabled;
             }
-            else if (element is HtmlInputElement)
+            else if (element is HtmlInputElement inputElement)
             {
-                return !((HtmlInputElement)element).IsDisabled;
+                return !inputElement.IsDisabled;
             }
-            else if (element is HtmlSelectElement)
+            else if (element is HtmlSelectElement selectElement)
             {
-                return !((HtmlSelectElement)element).IsDisabled;
+                return !selectElement.IsDisabled;
             }
-            else if (element is HtmlTextAreaElement)
+            else if (element is HtmlTextAreaElement textAreaElement)
             {
-                return !((HtmlTextAreaElement)element).IsDisabled;
+                return !textAreaElement.IsDisabled;
             }
-            else if (element is HtmlOptionElement)
+            else if (element is HtmlOptionElement optionElement)
             {
-                return !((HtmlOptionElement)element).IsDisabled;
+                return !optionElement.IsDisabled;
             }
             else if (element is HtmlOptionsGroupElement || element is HtmlMenuItemElement || element is HtmlFieldSetElement)
             {
@@ -350,25 +347,25 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is currently disabled, otherwise false.</returns>
         public static Boolean IsDisabled(this IElement element)
         {
-            if (element is HtmlButtonElement)
+            if (element is HtmlButtonElement buttonElement)
             {
-                return ((HtmlButtonElement)element).IsDisabled;
+                return buttonElement.IsDisabled;
             }
-            else if (element is HtmlInputElement)
+            else if (element is HtmlInputElement inputElement)
             {
-                return ((HtmlInputElement)element).IsDisabled;
+                return inputElement.IsDisabled;
             }
-            else if (element is HtmlSelectElement)
+            else if (element is HtmlSelectElement selectElement)
             {
-                return ((HtmlSelectElement)element).IsDisabled;
+                return selectElement.IsDisabled;
             }
-            else if (element is HtmlTextAreaElement)
+            else if (element is HtmlTextAreaElement textAreaElement)
             {
-                return ((HtmlTextAreaElement)element).IsDisabled;
+                return textAreaElement.IsDisabled;
             }
-            else if (element is HtmlOptionElement)
+            else if (element is HtmlOptionElement optionElement)
             {
-                return ((HtmlOptionElement)element).IsDisabled;
+                return optionElement.IsDisabled;
             }
             else if (element is HtmlOptionsGroupElement || element is HtmlMenuItemElement || element is HtmlFieldSetElement)
             {
@@ -386,10 +383,9 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is currently in its default state, otherwise false.</returns>
         public static Boolean IsDefault(this IElement element)
         {
-            if (element is HtmlButtonElement)
+            if (element is HtmlButtonElement buttonElement)
             {
-                var bt = (HtmlButtonElement)element;
-                var form = bt.Form;
+                var form = buttonElement.Form;
 
                 //TODO Check if button is form def. button
                 if (form != null)
@@ -397,14 +393,13 @@ namespace AngleSharp.Dom
                     return true;
                 }
             }
-            else if (element is HtmlInputElement)
+            else if (element is HtmlInputElement inputElement)
             {
-                var input = (HtmlInputElement)element;
-                var type = input.Type;
+                var type = inputElement.Type;
 
                 if (type == InputTypeNames.Submit || type == InputTypeNames.Image)
                 {
-                    var form = input.Form;
+                    var form = inputElement.Form;
 
                     //TODO Check if input is form def. button
                     if (form != null)
@@ -445,23 +440,20 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is currently checked, otherwise false.</returns>
         public static Boolean IsChecked(this IElement element)
         {
-            if (element is HtmlInputElement)
+            if (element is HtmlInputElement input)
             {
-                var input = (HtmlInputElement)element;
                 var type = input.Type;
                 var canBeChecked = type.IsOneOf(InputTypeNames.Checkbox, InputTypeNames.Radio);
                 return canBeChecked && input.IsChecked;
             }
-            else if (element is HtmlMenuItemElement)
+            else if (element is HtmlMenuItemElement menuItem)
             {
-                var menuItem = (HtmlMenuItemElement)element;
                 var type = menuItem.Type;
                 var canBeChecked = type.IsOneOf(InputTypeNames.Checkbox, InputTypeNames.Radio);
                 return canBeChecked && menuItem.IsChecked;
             }
-            else if (element is HtmlOptionElement)
+            else if (element is HtmlOptionElement option)
             {
-                var option = ((HtmlOptionElement)element);
                 return option.IsSelected;
             }
 
@@ -475,9 +467,8 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is currently indeterminate, otherwise false.</returns>
         public static Boolean IsIndeterminate(this IElement element)
         {
-            if (element is HtmlInputElement)
+            if (element is HtmlInputElement input)
             {
-                var input = (HtmlInputElement)element;
                 var isCheckbox = input.Type.Is(InputTypeNames.Checkbox);
                 return isCheckbox && input.IsIndeterminate;
             }
@@ -514,23 +505,20 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is currently unchecked, otherwise false.</returns>
         public static Boolean IsUnchecked(this IElement element)
         {
-            if (element is HtmlInputElement)
+            if (element is HtmlInputElement input)
             {
-                var input = (HtmlInputElement)element;
                 var type = input.Type;
                 var canBeChecked = type.IsOneOf(InputTypeNames.Checkbox, InputTypeNames.Radio);
                 return canBeChecked && !input.IsChecked;
             }
-            else if (element is HtmlMenuItemElement)
+            else if (element is HtmlMenuItemElement menuItem)
             {
-                var menuItem = (HtmlMenuItemElement)element;
                 var type = menuItem.Type;
                 var canBeChecked = type.IsOneOf(InputTypeNames.Checkbox, InputTypeNames.Radio);
                 return canBeChecked && !menuItem.IsChecked;
             }
-            else if (element is HtmlOptionElement)
+            else if (element is HtmlOptionElement option)
             {
-                var option = (HtmlOptionElement)element;
                 return !option.IsSelected;
             }
 
@@ -544,39 +532,33 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is currently active, otherwise false.</returns>
         public static Boolean IsActive(this IElement element)
         {
-            if (element is HtmlAnchorElement)
+            if (element is HtmlAnchorElement anchor)
             {
-                var anchor = (HtmlAnchorElement)element;
                 var href = element.GetAttribute(null, AttributeNames.Href);
                 return !String.IsNullOrEmpty(href) && anchor.IsActive;
             }
-            else if (element is HtmlAreaElement)
+            else if (element is HtmlAreaElement area)
             {
-                var area = (HtmlAreaElement)element;
                 var href = element.GetAttribute(null, AttributeNames.Href);
                 return !String.IsNullOrEmpty(href) && area.IsActive;
             }
-            else if (element is HtmlLinkElement)
+            else if (element is HtmlLinkElement link)
             {
-                var link = (HtmlLinkElement)element;
                 var href = element.GetAttribute(null, AttributeNames.Href);
                 return !String.IsNullOrEmpty(href) && link.IsActive;
             }
-            else if (element is HtmlButtonElement)
+            else if (element is HtmlButtonElement button)
             {
-                var button = (HtmlButtonElement)element;
                 return !button.IsDisabled && button.IsActive;
             }
-            else if (element is HtmlInputElement)
+            else if (element is HtmlInputElement input)
             {
-                var input = (HtmlInputElement)element;
                 var type = input.Type;
                 var canBeSubmitted = type.IsOneOf(InputTypeNames.Submit, InputTypeNames.Image, InputTypeNames.Reset, InputTypeNames.Button);
                 return canBeSubmitted && input.IsActive;
             }
-            else if (element is HtmlMenuItemElement)
+            else if (element is HtmlMenuItemElement menuItem)
             {
-                var menuItem = (HtmlMenuItemElement)element;
                 return !menuItem.IsDisabled && menuItem.IsActive;
             }
 
@@ -590,22 +572,19 @@ namespace AngleSharp.Dom
         /// <returns>True if the element has been visited, otherwise false.</returns>
         public static Boolean IsVisited(this IElement element)
         {
-            if (element is HtmlAnchorElement)
+            if (element is HtmlAnchorElement anchor)
             {
                 var href = element.GetAttribute(null, AttributeNames.Href);
-                var anchor = (HtmlAnchorElement)element;
                 return !String.IsNullOrEmpty(href) && anchor.IsVisited;
             }
-            else if (element is HtmlAreaElement)
+            else if (element is HtmlAreaElement area)
             {
                 var href = element.GetAttribute(null, AttributeNames.Href);
-                var area = (HtmlAreaElement)element;
                 return !String.IsNullOrEmpty(href) && area.IsVisited;
             }
-            else if (element is HtmlLinkElement)
+            else if (element is HtmlLinkElement link)
             {
                 var href = element.GetAttribute(null, AttributeNames.Href);
-                var link = (HtmlLinkElement)element;
                 return !String.IsNullOrEmpty(href) && link.IsVisited;
             }
 
@@ -619,22 +598,19 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is a link, otherwise false.</returns>
         public static Boolean IsLink(this IElement element)
         {
-            if (element is HtmlAnchorElement)
+            if (element is HtmlAnchorElement anchor)
             {
                 var href = element.GetAttribute(null, AttributeNames.Href);
-                var anchor = (HtmlAnchorElement)element;
                 return !String.IsNullOrEmpty(href) && !anchor.IsVisited;
             }
-            else if (element is HtmlAreaElement)
+            else if (element is HtmlAreaElement area)
             {
                 var href = element.GetAttribute(null, AttributeNames.Href);
-                var area = (HtmlAreaElement)element;
                 return !String.IsNullOrEmpty(href) && !area.IsVisited;
             }
-            else if (element is HtmlLinkElement)
+            else if (element is HtmlLinkElement link)
             {
                 var href = element.GetAttribute(null, AttributeNames.Href);
-                var link = (HtmlLinkElement)element;
                 return !String.IsNullOrEmpty(href) && !link.IsVisited;
             }
 
@@ -646,10 +622,7 @@ namespace AngleSharp.Dom
         /// </summary>
         /// <param name="element">The element to check.</param>
         /// <returns>True if the element hosts a shadow tree.</returns>
-        public static Boolean IsShadow(this IElement element)
-        {
-            return element?.ShadowRoot != null;
-        }
+        public static Boolean IsShadow(this IElement element) => element?.ShadowRoot != null;
 
         /// <summary>
         /// Checks if the element is only optional and not required.
@@ -658,19 +631,16 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is optional, otherwise false.</returns>
         public static Boolean IsOptional(this IElement element)
         {
-            if (element is HtmlInputElement)
+            if (element is HtmlInputElement input)
             {
-                var input = (HtmlInputElement)element;
                 return !input.IsRequired;
             }
-            else if (element is HtmlSelectElement)
+            else if (element is HtmlSelectElement select)
             {
-                var select = (HtmlSelectElement)element;
                 return !select.IsRequired;
             }
-            else if (element is HtmlTextAreaElement)
+            else if (element is HtmlTextAreaElement area)
             {
-                var area = (HtmlTextAreaElement)element;
                 return !area.IsRequired;
             }
 
@@ -684,19 +654,16 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is required, otherwise false.</returns>
         public static Boolean IsRequired(this IElement element)
         {
-            if (element is HtmlInputElement)
+            if (element is HtmlInputElement input)
             {
-                var input = (HtmlInputElement)element;
                 return input.IsRequired;
             }
-            else if (element is HtmlSelectElement)
+            else if (element is HtmlSelectElement select)
             {
-                var select = (HtmlSelectElement)element;
                 return select.IsRequired;
             }
-            else if (element is HtmlTextAreaElement)
+            else if (element is HtmlTextAreaElement textArea)
             {
-                var textArea = (HtmlTextAreaElement)element;
                 return textArea.IsRequired;
             }
 
@@ -710,14 +677,12 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is invalid, otherwise false.</returns>
         public static Boolean IsInvalid(this IElement element)
         {
-            if (element is IValidation)
+            if (element is IValidation validator)
             {
-                var validator = (IValidation)element;
                 return !validator.CheckValidity();
             }
-            else if (element is HtmlFormElement)
+            else if (element is HtmlFormElement form)
             {
-                var form = (HtmlFormElement)element;
                 return !form.CheckValidity();
             }
 
@@ -731,14 +696,12 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is valid, otherwise false.</returns>
         public static Boolean IsValid(this IElement element)
         {
-            if (element is IValidation)
+            if (element is IValidation validator)
             {
-                var validator = (IValidation)element;
                 return validator.CheckValidity();
             }
-            else if (element is HtmlFormElement)
+            else if (element is HtmlFormElement form)
             {
-                var form = (HtmlFormElement)element;
                 return form.CheckValidity();
             }
 
@@ -752,19 +715,16 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is readonly, otherwise false.</returns>
         public static Boolean IsReadOnly(this IElement element)
         {
-            if (element is HtmlInputElement)
+            if (element is HtmlInputElement input)
             {
-                var input = (HtmlInputElement)element;
                 return !input.IsMutable;
             }
-            else if (element is HtmlTextAreaElement)
+            else if (element is HtmlTextAreaElement textArea)
             {
-                var textArea = (HtmlTextAreaElement)element;
                 return !textArea.IsMutable;
             }
-            else if (element is IHtmlElement)
+            else if (element is IHtmlElement general)
             {
-                var general = (IHtmlElement)element;
                 return !general.IsContentEditable;
             }
 
@@ -778,19 +738,16 @@ namespace AngleSharp.Dom
         /// <returns>True if the element can be edited, otherwise false.</returns>
         public static Boolean IsEditable(this IElement element)
         {
-            if (element is HtmlInputElement)
+            if (element is HtmlInputElement input)
             {
-                var input = (HtmlInputElement)element;
                 return input.IsMutable;
             }
-            else if (element is HtmlTextAreaElement)
+            else if (element is HtmlTextAreaElement textArea)
             {
-                var textArea = (HtmlTextAreaElement)element;
                 return textArea.IsMutable;
             }
-            else if (element is IHtmlElement)
+            else if (element is IHtmlElement general)
             {
-                var general = (IHtmlElement)element;
                 return general.IsContentEditable;
             }
 
@@ -939,10 +896,7 @@ namespace AngleSharp.Dom
         /// <param name="attributeName">The name of the attribute.</param>
         /// <returns>The attributes' values.</returns>
         public static IEnumerable<String> Attr<T>(this T elements, String attributeName)
-            where T : IEnumerable<IElement>
-        {
-            return elements.Select(m => m.GetAttribute(attributeName));
-        }
+            where T : IEnumerable<IElement> => elements.Select(m => m.GetAttribute(attributeName));
 
         /// <summary>
         /// Clears the attributes of the given element.
@@ -1280,12 +1234,9 @@ namespace AngleSharp.Dom
         /// <param name="html">The HTML code that generates the tree.</param>
         /// <returns>The unchanged collection.</returns>
         public static T WrapInner<T>(this T elements, String html)
-            where T : IEnumerable<IElement>
+            where T : class, IEnumerable<IElement>
         {
-            if (elements == null)
-            {
-                throw new ArgumentNullException(nameof(elements));
-            }
+            elements = elements ?? throw new ArgumentNullException(nameof(elements));
 
             foreach (var element in elements)
             {
@@ -1314,13 +1265,9 @@ namespace AngleSharp.Dom
         /// <param name="html">The HTML code that generates the tree.</param>
         /// <returns>The unchanged collection.</returns>
         public static T WrapAll<T>(this T elements, String html)
-            where T : IEnumerable<IElement>
+            where T : class, IEnumerable<IElement>
         {
-            if (elements == null)
-            {
-                throw new ArgumentNullException(nameof(elements));
-            }
-
+            elements = elements ?? throw new ArgumentNullException(nameof(elements));
             var element = elements.FirstOrDefault();
 
             if (element != null)
@@ -1346,10 +1293,7 @@ namespace AngleSharp.Dom
         /// <param name="elements">The elements to include.</param>
         /// <returns>The created collection.</returns>
         public static IHtmlCollection<TElement> ToCollection<TElement>(this IEnumerable<TElement> elements)
-            where TElement : class, IElement
-        {
-            return new HtmlCollection<TElement>(elements);
-        }
+            where TElement : class, IElement => new HtmlCollection<TElement>(elements);
 
         /// <summary>
         /// Navigates to the hyper reference given by the provided element
@@ -1359,10 +1303,7 @@ namespace AngleSharp.Dom
         /// <param name="element">The element of navigation.</param>
         /// <returns>The task eventually resulting in the response.</returns>
         public static Task<IDocument> NavigateAsync<TElement>(this TElement element)
-            where TElement : class, IUrlUtilities, IElement
-        {
-            return element.NavigateAsync(CancellationToken.None);
-        }
+            where TElement : class, IUrlUtilities, IElement => element.NavigateAsync(CancellationToken.None);
 
         /// <summary>
         /// Navigates to the hyper reference given by the provided element.
@@ -1455,10 +1396,7 @@ namespace AngleSharp.Dom
         /// <param name="element">The element to host the attribute.</param>
         /// <param name="name">The name of the attribute.</param>
         /// <returns>The attribute's value, if any.</returns>
-        internal static String GetOwnAttribute(this Element element, String name)
-        {
-            return element.Attributes.GetNamedItem(null, name)?.Value;
-        }
+        internal static String GetOwnAttribute(this Element element, String name) => element.Attributes.GetNamedItem(null, name)?.Value;
 
         /// <summary>
         /// Faster way of checking for a (known) attribute.
@@ -1466,10 +1404,7 @@ namespace AngleSharp.Dom
         /// <param name="element">The element to host the attribute.</param>
         /// <param name="name">The name of the attribute.</param>
         /// <returns>True if the attribute exists, otherwise false.</returns>
-        internal static Boolean HasOwnAttribute(this Element element, String name)
-        {
-            return element.Attributes.GetNamedItem(null, name) != null;
-        }
+        internal static Boolean HasOwnAttribute(this Element element, String name) => element.Attributes.GetNamedItem(null, name) != null;
 
         /// <summary>
         /// Easy way of getting the current absolute url from attributes.
@@ -1521,10 +1456,7 @@ namespace AngleSharp.Dom
         /// <param name="name">The name of the attribute.</param>
         /// <param name="value">The attribute's value.</param>
         /// <param name="suppressCallbacks">Flag to suppress callbacks.</param>
-        internal static void SetOwnAttribute(this Element element, String name, String value, Boolean suppressCallbacks = false)
-        {
-            element.Attributes.SetNamedItemWithNamespaceUri(new Attr(name, value), suppressCallbacks);
-        }
+        internal static void SetOwnAttribute(this Element element, String name, String value, Boolean suppressCallbacks = false) => element.Attributes.SetNamedItemWithNamespaceUri(new Attr(name, value), suppressCallbacks);
 
         private static IDocumentFragment CreateFragment(this IElement context, String html)
         {
@@ -1556,12 +1488,12 @@ namespace AngleSharp.Dom
         /// <summary>
         /// Creates a unique selector path used to locate the element in the DOM.
         /// </summary>
-        /// <param name="node">The starting node to create the selector path from.</param>
+        /// <param name="element">The starting node to create the selector path from.</param>
         /// <returns>The unique selector path for this element.</returns>
-        public static string GetSelector(this IElement node)
+        public static String GetSelector(this IElement element)
         {
             // Initialize path
-            var path = string.Empty;
+            var path = String.Empty;
 
             // If the current node is having an unique id property
             var hasId = false;
@@ -1569,36 +1501,44 @@ namespace AngleSharp.Dom
             do
             {
                 // Set if node has id attribute set...
-                hasId = !string.IsNullOrEmpty(node.Id);
+                hasId = !String.IsNullOrEmpty(element.Id);
 
                 // Get parent element of the node
-                var parent = node.ParentElement;
+                var parent = element.ParentElement;
 
                 // Always lowercase node name in the path
-                var name = node.NodeName.ToLowerInvariant();
+                var name = element.LocalName;
 
                 // If node has id attribute...
                 if (hasId)
                 {
                     // Id is unique in the DOM, so we can use it to locate the element and skip other parents
-                    name = "#" + node.Id;
+                    name = "#" + element.Id;
                 }
                 // If node has siblings of the same type...
-                else if (parent != null && !node.IsOnlyOfType())
+                else if (parent != null && !element.IsOnlyOfType())
                 {
                     // Get node index in the parent node tree
-                    var index = parent.Children.Where(_ => _.GetType() == node.GetType()).Index(node);
+                    var index = parent.Children.Where(_ => _.GetType() == element.GetType()).Index(element);
 
                     // Append nth child selector
                     name += $":nth-child({index + 1})";
                 }
 
-                // Set current parent
-                node = parent;
-
                 // Recreate selector path
-                path = name + (!string.IsNullOrEmpty(path) ? ">" + path : "");
-            } while (node?.ParentElement != null && !hasId);
+                if (!String.IsNullOrEmpty(path))
+                {
+                    path = $"{name}>{path}";
+                }
+                else
+                {
+                    path = name;
+                }
+
+                // Set current parent
+                element = parent;
+            }
+            while (element?.ParentElement != null && !hasId);
 
             // Return generated selector
             return path;