+++ src/AngleSharp/Html/HtmlMarkupFormatter.cs
@@ -20,10 +20,7 @@ namespace AngleSharp.Html
 
         #region Methods
 
-        String IMarkupFormatter.Comment(IComment comment)
-        {
-            return String.Concat("<!--", comment.Data, "-->");
-        }
+        String IMarkupFormatter.Comment(IComment comment) => String.Concat("<!--", comment.Data, "-->");
 
         String IMarkupFormatter.Doctype(IDocumentType doctype)
         {
@@ -163,10 +160,7 @@ namespace AngleSharp.Html
             return $" PUBLIC \"{publicId}\" \"{systemId}\"";
         }
 
-        private static String XmlNamespaceLocalName(String name)
-        {
-            return name != NamespaceNames.XmlNsPrefix ? String.Concat(NamespaceNames.XmlNsPrefix, ":", name) : name;
-        }
+        private static String XmlNamespaceLocalName(String name) => name != NamespaceNames.XmlNsPrefix ? String.Concat(NamespaceNames.XmlNsPrefix, ":", name) : name;
 
         #endregion
     }