+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 0.12.0
 
-Released on Thursday, May 5 2019.
+Released on Thursday, May 2 2019.
 
+- Added ability to create a selector for an element (#784)
 - Enhanced documentation (#776, #774, #771)
 - Fixed DOM name attribute in `AdjacentPosition` (#775)
 - Fixed bug regarding `set-cookie` header (#768)